Add files via upload
Stroke is ranked second leading cause of death according to the World Health Organisation{WHO}. It is a major public health concern affecting millions of people worldwide and it is responsible for approximately 11% of deaths. Stroke occurs due to a decrease or blockage in the brain’s blood supply when a blood vessel that carries oxygen and nutrient to the brain is either blocked by a clot or burst, when that happens part of the brain cannot get the blood and oxygen it needs which result to the death of the brain.
It is also said to be the common cause of long term disability. This remains a health burden for the individuals and the National Health care system which should not be overlooked. Stroke awareness is very important to aid its prevention.
This dataset contains records from both ischemic and hemorrhagic strokes which will assist in predicting the likelihood of a patient having a stroke based on  results obtained from the analysis, timely detection and prevention has become very essential to avoid its adverse consequences. 
For the purpose of this report, the target audience will be stroke enthusiasts. This is to ensure that they have the right information and resources required to educate the public.